+++ draft-packages/title-block-zen/KaizenDraft/TitleBlockZen/index.ts
@@ -1,0 +1,2 @@
+export { default as TitleBlockZen } from "./TitleBlockZen"
+export { NavigationTab } from "./TitleBlockZen"