+++ draft-packages/events/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.35](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-events@1.2.34...@kaizen/draft-events@1.2.35) (2020-06-16)
+
+**Note:** Version bump only for package @kaizen/draft-events
+
+
+
+
+
 ## [1.2.34](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-events@1.2.33...@kaizen/draft-events@1.2.34) (2020-06-12)
 
 **Note:** Version bump only for package @kaizen/draft-events